pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "player-profiles"
version = "0.1.0"
description = "Hierarchical Bayesian MLB player projection and pitch-type matchup system"
requires-python = ">=3.11"
dependencies = [
    "pymc>=5.10",
    "arviz>=0.17",
    "pandas>=2.1",
    "numpy>=1.25",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "scikit-learn>=1.3",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.30",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ipykernel>=6.25",
    "jupyter>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: slower end-to-end tests that sample real PyMC models on synthetic data",
]